pyproject.toml
[tool.poetry]
name = "birdwatcher"
version = "0.1.0"
description = ""
authors = ["Sean Moss <smoss117@gmail.com>"]

[tool.poetry.dependencies]
python = "3.10"
tensorflow-gpu = "^2.9.1"
tensorflow-hub = "^0.12.0"
pypng = "^0.20220715.0"
mypy = "^0.971"
keras = "^2.9.0"
pandas = "^1.4.3"
Pillow = "^9.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
